requirements.txt
requests>=2.28.0
pyyaml>=6.0
browser-cookie3>=0.19.0
schedule>=1.2.0
selenium>=4.6.0
